moves dto into shared
