+++ openvair/common/serialization/base_serializer.py
@@ -23,6 +23,7 @@ Dependencies:
 from typing import Any, Dict, Type, Generic, TypeVar, ClassVar, cast
 
 from pydantic import BaseModel
+from sqlalchemy import inspect
 from sqlalchemy.orm import DeclarativeBase
 
 from openvair.common.serialization.serializer_hub import SerializerHub
@@ -100,15 +101,46 @@ class BaseSerializer(Generic[DTO, ORM]):
         Example:
             orm = TemplateSerializer.to_orm(dto)
         """
-        fields = cls.orm_class.__table__.columns.keys()
         dto_data = dto_obj.model_dump()
+        inspected = inspect(cls.orm_class)
+        orm_fields = {col.key for col in inspected.mapper.column_attrs}
+
         nested_data: Dict[str, Any] = {}
         for field, value in dto_data.items():
             if field in cls.nested_serializers and isinstance(value, list):
                 nested = cls.nested_serializers[field]
                 nested_data[field] = [nested.to_orm(item) for item in value]
-        filtered = {k: v for k, v in dto_data.items() if k in fields}
+
+        filtered = {k: v for k, v in dto_data.items() if k in orm_fields}
         orm = cls.orm_class(**filtered)
+
         for key, val in nested_data.items():
             setattr(orm, key, val)
+
         return cast(ORM, orm)
+
+    @classmethod
+    def from_dict(cls, data: Dict[str, Any]) -> ORM:
+        """Convert raw dictionary to ORM model via DTO.
+
+        Args:
+            data (Dict[str, Any]): Input data as dictionary.
+
+        Returns:
+            ORM: ORM object constructed from the data.
+        """
+        dto = cls.dto_class.model_validate(data)
+        return cls.to_orm(dto)
+
+    @classmethod
+    def to_dict(cls, orm_obj: ORM) -> Dict[str, Any]:
+        """Convert ORM object to dictionary via DTO.
+
+        Args:
+            orm_obj (ORM): ORM model instance.
+
+        Returns:
+            Dict[str, Any]: Dictionary representation of the object.
+        """
+        dto = cls.to_dto(orm_obj)
+        return dto.model_dump(mode='json')
