+++ openvair/modules/template/adapters/serializer.py
@@ -7,8 +7,8 @@ Classes:
     - TemplateSerializer: Converts Template <-> TemplateDTO
 """
 
+from openvair.modules.template.shared.dto import TemplateDTO
 from openvair.modules.template.adapters.orm import Template as TemplateORM
-from openvair.modules.template.service_layer.dtos import TemplateDTO
 from openvair.common.serialization.base_serializer import BaseSerializer
